進行バーの実装/HPゲージの表示位置の調整
