+++ Assets/Scripts/ProgressBar.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 5fb9e85df547d1d4189971d13da6842d
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
